Inline auth logic into controllers and remove authService
