+++ src/controllers/callbackController.ts
@@ -1,6 +1,18 @@
 import { Request, Response } from "express";
-import { fetchUserData } from "../services/authService";
+import axios from "axios";
 import appConfig from "../config/appConfig";
+
+const constructTokenParams = (authCode: string) => {
+  const tokenParams = new URLSearchParams();
+  tokenParams.append("grant_type", "authorization_code");
+  tokenParams.append("code", authCode);
+  tokenParams.append(
+    "redirect_uri",
+    `${appConfig.baseUrl}/api/auth/callback`
+  );
+  tokenParams.append("client_id", appConfig.wcaClientId);
+  return tokenParams;
+};
 
 const callback = async (req: Request, res: Response): Promise<void> => {
   const code = req.query.code;
@@ -11,9 +23,31 @@ const callback = async (req: Request, res: Response): Promise<void> => {
     return;
   }
   try {
-    const user = await fetchUserData(code);
+    const tokenUrl = new URL(`${appConfig.wcaUrl}/oauth/token`);
+    const tokenParams = constructTokenParams(code);
+    const basicAuthToken = Buffer.from(
+      `${appConfig.wcaClientId}:${appConfig.wcaClientSecret}`
+    ).toString("base64");
+    const tokenResponse = await axios.post(
+      tokenUrl.toString(),
+      tokenParams.toString(),
+      {
+        headers: {
+          "Content-Type": "application/x-www-form-urlencoded",
+          Authorization: `Basic ${basicAuthToken}`,
+        },
+      }
+    );
+
+    const { access_token } = tokenResponse.data;
+
+    const userUrl = new URL(`${appConfig.wcaUrl}/api/v0/me`);
+    const userResponse = await axios.get(userUrl.toString(), {
+      headers: { Authorization: `Bearer ${access_token}` },
+    });
+
     if (process.env.NODE_ENV !== "production") {
-      console.log(user);
+      console.log(userResponse.data);
     }
     res.redirect(`${appConfig.frontendUrl}`);
   } catch (error) {
